+++ src/router/modules/demo/barcode.ts
@@ -5,7 +5,7 @@ import type { AppRouteRecordRaw } from '@/router/types'
 
 const barcode: AppRouteRecordRaw = {
   path: 'barcode',
-  component: () => import('@/views/demo/BarcodeDemo'),
+  component: () => import('@/views/demo/barcode-demo'),
   meta: {
     i18nKey: t('menu.Barcode'),
     icon: 'other',